+++ Libplanet.Tests/Action/ActionEvaluatorTest.cs
@@ -256,7 +256,7 @@ namespace Libplanet.Tests.Action
 
             Assert.Throws<OutOfMemoryException>(
                 () => actionEvaluator.EvaluateTx(
-                    blockHeader: block,
+                    block: block,
                     tx: tx,
                     previousState: previousState).ToList());
             Assert.Throws<OutOfMemoryException>(
@@ -577,7 +577,7 @@ namespace Libplanet.Tests.Action
 
             IWorld previousState = actionEvaluator.PrepareInitialDelta(null);
             var evaluations = actionEvaluator.EvaluateTx(
-                blockHeader: block,
+                block: block,
                 tx: tx,
                 previousState: previousState).ToImmutableArray();
 
@@ -648,7 +648,7 @@ namespace Libplanet.Tests.Action
 
             previousState = actionEvaluator.PrepareInitialDelta(null);
             IWorld delta = actionEvaluator.EvaluateTx(
-                blockHeader: block,
+                block: block,
                 tx: tx,
                 previousState: previousState).Last().OutputState;
             Assert.Empty(evaluations[3].OutputState.Trie.Diff(delta.Trie));
@@ -684,7 +684,7 @@ namespace Libplanet.Tests.Action
                 transactions: txs).Propose();
             IWorld previousState = actionEvaluator.PrepareInitialDelta(null);
             var nextState = actionEvaluator.EvaluateTx(
-                blockHeader: block,
+                block: block,
                 tx: tx,
                 previousState: previousState).Last().OutputState;
 
@@ -706,7 +706,7 @@ namespace Libplanet.Tests.Action
             Block blockA = fx.Propose();
             fx.Append(blockA);
             ActionEvaluation[] evalsA = ActionEvaluator.EvaluateActions(
-                blockHeader: blockA,
+                block: blockA,
                 tx: txA,
                 previousState: fx.CreateWorld(blockA.PreviousHash),
                 actions: txA.Actions
@@ -719,6 +719,7 @@ namespace Libplanet.Tests.Action
                 new Integer(15),
                 evalsA.Last().OutputState
                     .GetAccount(ReservedAddresses.LegacyAccount).GetState(txA.Signer));
+            Assert.All(evalsA, eval => Assert.Empty(eval.InputContext.Txs));
 
             for (int i = 0; i < evalsA.Length; i++)
             {
@@ -761,7 +762,7 @@ namespace Libplanet.Tests.Action
             Block blockB = fx.Propose();
             fx.Append(blockB);
             ActionEvaluation[] evalsB = ActionEvaluator.EvaluateActions(
-                blockHeader: blockB,
+                block: blockB,
                 tx: txB,
                 previousState: fx.CreateWorld(blockB.PreviousHash),
                 actions: txB.Actions
@@ -774,6 +775,7 @@ namespace Libplanet.Tests.Action
                 new Integer(6),
                 evalsB.Last().OutputState
                     .GetAccount(ReservedAddresses.LegacyAccount).GetState(txB.Signer));
+            Assert.All(evalsB, eval => Assert.Empty(eval.InputContext.Txs));
 
             for (int i = 0; i < evalsB.Length; i++)
             {
@@ -840,6 +842,7 @@ namespace Libplanet.Tests.Action
                 (Integer)evaluation.OutputState
                     .GetAccount(ReservedAddresses.LegacyAccount).GetState(genesis.Miner));
             Assert.True(evaluation.InputContext.BlockAction);
+            Assert.Equal(genesis.Transactions, evaluation.InputContext.Txs);
 
             previousState = evaluation.OutputState;
             evaluation = actionEvaluator.EvaluatePolicyBlockAction(block, previousState);
@@ -850,6 +853,7 @@ namespace Libplanet.Tests.Action
                 (Integer)evaluation.OutputState
                     .GetAccount(ReservedAddresses.LegacyAccount).GetState(block.Miner));
             Assert.True(evaluation.InputContext.BlockAction);
+            Assert.Equal(block.Transactions, evaluation.InputContext.Txs);
 
             chain.Append(block, CreateBlockCommit(block), render: true);
             previousState = actionEvaluator.PrepareInitialDelta(genesis.StateRootHash);
@@ -863,6 +867,7 @@ namespace Libplanet.Tests.Action
                 (Integer)2,
                 (Integer)evaluation.OutputState
                     .GetAccount(ReservedAddresses.LegacyAccount).GetState(block.Miner));
+            Assert.Equal(block.Transactions, evaluation.InputContext.Txs);
         }
 
         [Theory]
@@ -1280,7 +1285,7 @@ namespace Libplanet.Tests.Action
 
             Block blockA = fx.Propose();
             ActionEvaluation[] evalsA = ActionEvaluator.EvaluateActions(
-                blockHeader: blockA,
+                block: blockA,
                 tx: txA,
                 previousState: fx.CreateWorld(blockA.PreviousHash),
                 actions: txA.Actions